+++ JotunnLib/Managers/GUIManager.cs
@@ -178,7 +178,7 @@ namespace Jotunn.Managers
         ///     Counter to track multiple block requests.
         /// </summary>
         private static int InputBlockRequests;
-
+        
         /// <summary>
         ///     Block all input except GUI
         /// </summary>
@@ -437,6 +437,8 @@ namespace Jotunn.Managers
             CustomGUIFront.layer = UILayer;
             CustomGUIFront.transform.SetParent(parent.transform, false);
             CustomGUIFront.transform.SetAsLastSibling();
+            CustomGUIFront.GetComponent<RectTransform>().anchorMin = Vector2.zero;
+            CustomGUIFront.GetComponent<RectTransform>().anchorMax = Vector2.one;
 
 #pragma warning disable CS0618 // Type or member is obsolete
             PixelFix = CustomGUIFront;
@@ -446,6 +448,8 @@ namespace Jotunn.Managers
             CustomGUIBack.layer = UILayer;
             CustomGUIBack.transform.SetParent(parent.transform, false);
             CustomGUIBack.transform.SetAsFirstSibling();
+            CustomGUIBack.GetComponent<RectTransform>().anchorMin = Vector2.zero;
+            CustomGUIBack.GetComponent<RectTransform>().anchorMax = Vector2.one;
 
 #pragma warning disable CS0612 // Type or member is obsolete
             InvokeOnPixelFixCreated();
@@ -1082,11 +1086,13 @@ namespace Jotunn.Managers
             // Optionally set width and height
             if (width > 0f)
             {
+                float contentWidth = width - 50f;
+                dropdown.SetWidth(width);
                 dropdownComponent.captionText.gameObject.SetMiddleLeft();
-                dropdownComponent.captionText.gameObject.SetWidth(width);
+                dropdownComponent.captionText.gameObject.SetWidth(contentWidth);
                 dropdownComponent.captionText.GetComponent<RectTransform>()
                     .anchoredPosition = new Vector3(10f, 0f);
-                dropdownComponent.itemText.gameObject.SetWidth(width);
+                dropdownComponent.itemText.gameObject.SetWidth(contentWidth);
             }
 
             if (height > 0f)